Fix .chart HOPO threshold not calculating correctly

+++ YARG.Core/MoonscraperChartParser/IO/Chart/ChartIOHelper.cs
@@ -108,7 +108,7 @@ namespace MoonscraperChartEditor.Song.IO
             // so we need to scale this factor to different resolutions (480 res = 162.5 threshold)
             // This extra tick is meant for some slight leniency, .mid has it too but it's applied
             // after factoring in the resolution there, not before
-            const int DEFAULT_RESOLUTION = 192;
+            const float DEFAULT_RESOLUTION = 192;
             const float HOPO_THRESHOLD_FACTOR = ((DEFAULT_RESOLUTION / 3) + 1) / DEFAULT_RESOLUTION;
             const float EIGHTHNOTE_HOPO_THRESHOLD_FACTOR = ((DEFAULT_RESOLUTION / 2) + 1) / DEFAULT_RESOLUTION;
